% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_position.R
\name{closest_gene}
\alias{closest_gene}
\title{Get closest gene bnumber}
\usage{
closest_gene(position, strand = "+-")
}
\arguments{
\item{position}{A numeric vector}

\item{strand}{A character in c("+", "-", "+-")}
}
\value{
A vector of characters of same size as input
}
\description{
Get closest gene bnumber
}
