% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coords.R
\name{convert_coords}
\alias{convert_coords}
\title{Convert genomic coordinates from 96.2 to 96.3}
\usage{
convert_coords(coords)
}
\arguments{
\item{coords}{A integer vector}
}
\value{
A integer vector of same size as \code{coords}
}
\description{
Convert genomic coordinates from 96.2 to 96.3
}
