% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene.R
\name{convert_gene}
\alias{convert_gene}
\title{Convert gene names, IDs etc to reference gene symbols or bnumbers}
\usage{
convert_gene(genes, to = "bnumber")
}
\arguments{
\item{genes}{A character vector}

\item{to}{A character string}
}
\value{
A character vector of same size as \code{genes}
}
\description{
Convert gene names, IDs etc to reference gene symbols or bnumbers
}
