% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tf.R
\name{convert_tf}
\alias{convert_tf}
\title{Get TF  names.}
\usage{
convert_tf(tfs, to = "name")
}
\arguments{
\item{tfs}{A character vector}

\item{to}{A character string}
}
\value{
A character vector of same size as \code{tfs}
}
\description{
Get TF  names.
}
