% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_all_protein_synonyms}
\alias{get_all_protein_synonyms}
\title{Get all genes names and bnumbers}
\usage{
get_all_protein_synonyms()
}
\value{
A character vector
}
\description{
Get all genes names and bnumbers
}
