% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfs.R
\name{get_confirmed_tfs}
\alias{get_confirmed_tfs}
\title{Get all confirmed TFs from RegulonDB}
\usage{
get_confirmed_tfs()
}
\value{
A character vector of TF names
}
\description{
Get all confirmed TFs from RegulonDB
}
