% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_downstream_gene}
\alias{get_downstream_gene}
\title{Get next genes bnumbers for a given vector of bnumbers, with respect to strand orientation}
\usage{
get_downstream_gene(list_genes)
}
\arguments{
\item{list_genes}{A character vector of valid bnumbers (previously checked/converted with convert_gene(to = "bnumber"))}
}
\value{
A vector of same size as `bnumbers
}
\description{
Get next genes bnumbers for a given vector of bnumbers, with respect to strand orientation
}
