% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene.R
\name{get_gene_bnumber}
\alias{get_gene_bnumber}
\title{Convert gene symbols to bnumbers}
\usage{
get_gene_bnumber(list_genes)
}
\arguments{
\item{list_genes}{A character vector}
}
\value{
A character vector of same size as \code{list_genes}
}
\description{
Convert gene symbols to bnumbers
}
