% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_gene_info}
\alias{get_gene_info}
\title{Get reference info from the master table for a given list of genes}
\usage{
get_gene_info(list_genes)
}
\arguments{
\item{list_genes}{A character vector of gene names, bnumbers, symbols}
}
\value{
A dataframe where ncol = 7 and nrow = the number of valid genes provided as input
}
\description{
Get reference info from the master table for a given list of genes
}
