% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_gene_length}
\alias{get_gene_length}
\title{Get gene nucleotide length}
\usage{
get_gene_length(list_genes)
}
\arguments{
\item{list_genes}{A character vector of gene names, bnumbers, symbols}
}
\value{
A integer vector of same size as \code{list_genes}
}
\description{
Get gene nucleotide length
}
