% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_gene_ref}
\alias{get_gene_ref}
\title{Get info from the master table for a given vector of bnumbers}
\usage{
get_gene_ref(list_genes)
}
\arguments{
\item{list_genes}{A character vector of gene names, bnumbers, symbols}
}
\value{
A dataframe where ncol = 5 and nrow = the number of valid genes provided as input
}
\description{
Get info from the master table for a given vector of bnumbers
}
