% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_coords.R
\name{get_gene_stop}
\alias{get_gene_stop}
\title{Get gene stop}
\usage{
get_gene_stop(list_genes)
}
\arguments{
\item{list_genes}{A character vector of bnumbers (previously checked/converted)}
}
\value{
A character vector of same size as \code{list_genes}
}
\description{
Get gene stop
}
