% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene.R
\name{get_gene_symbol}
\alias{get_gene_symbol}
\title{Convert any name or id to gene symbols \link{deprecated}}
\usage{
get_gene_symbol(list_genes)
}
\arguments{
\item{list_genes}{A character vector}
}
\value{
A character vector of same size as \code{list_genes}
}
\description{
Convert any name or id to gene symbols \link{deprecated}
}
