% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_gene_synonyms}
\alias{get_gene_synonyms}
\title{Get gene synonyms from the master table for a given vector of genes}
\usage{
get_gene_synonyms(list_genes)
}
\arguments{
\item{list_genes}{A character vector of gene names, bnumbers, symbols}
}
\value{
A character vector of same size as \code{list_genes} with comma-separated synonyms
}
\description{
Get gene synonyms from the master table for a given vector of genes
}
