% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_next_gene}
\alias{get_next_gene}
\title{Get next genes bnumbers for a given vector of bnumbers, with respect to strand orientation}
\usage{
get_next_gene(list_genes)
}
\arguments{
\item{list_genes}{A character vector of gene names, bnumbers, symbols}
}
\value{
A vector of same size as `bnumbers
}
\description{
Get next genes bnumbers for a given vector of bnumbers, with respect to strand orientation
}
