% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfs.R
\name{get_predicted_tfs}
\alias{get_predicted_tfs}
\title{Get predicted TFs from the literature (Perez-Rueda 2015, Flores-Bautista 2020, Kim 2020)}
\usage{
get_predicted_tfs()
}
\value{
A character vector of TF names
}
\description{
Get predicted TFs from the literature (Perez-Rueda 2015, Flores-Bautista 2020, Kim 2020)
}
