% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_previous_gene}
\alias{get_previous_gene}
\title{Get previous genes bnumbers for a given vector of genes, with respect to strand orientation}
\usage{
get_previous_gene(list_genes)
}
\arguments{
\item{list_genes}{A character vector of gene names, bnumbers, symbols}
}
\value{
A vector of same size as `bnumbers
}
\description{
Get previous genes bnumbers for a given vector of genes, with respect to strand orientation
}
