% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_target_genes}
\alias{get_target_genes}
\title{Get potentially TF-regulated genes}
\usage{
get_target_genes(source = "all")
}
\arguments{
\item{source}{Optional, one of c("all", "zika", "regulondb")}
}
\value{
A character vector
}
\description{
Get potentially TF-regulated genes
}
