% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_protein.R
\name{get_tf_bnumber}
\alias{get_tf_bnumber}
\title{Convert TF protein names into bnumbers}
\usage{
get_tf_bnumber(list_proteins)
}
\arguments{
\item{list_proteins}{A character vector}
}
\value{
A character vector of same size as \code{list_proteins}
}
\description{
Convert TF protein names into bnumbers
}
