% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_info.R
\name{get_tf_effect}
\alias{get_tf_effect}
\title{Get conformation_type info for a given TF or list of TFs}
\usage{
get_tf_effect(list_tfs)
}
\value{
A character vector of same size as \code{list_tfs}
}
\description{
Get conformation_type info for a given TF or list of TFs
}
