% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_info.R
\name{get_tf_family}
\alias{get_tf_family}
\title{Get TF family}
\usage{
get_tf_family(list_tfs)
}
\arguments{
\item{list_tfs}{}
}
\value{
A character vector of same size as \code{list_tfs}
}
\description{
Get TF family
}
