% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_tf_genes}
\alias{get_tf_genes}
\title{Get TF-coding genes from RegulonDB (does not include putative TFs from other sources; check out get_tfs() functions instead).}
\usage{
get_tf_genes()
}
\value{
A character vector
}
\description{
Get TF-coding genes from RegulonDB (does not include putative TFs from other sources; check out get_tfs() functions instead).
}
