% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tf.R
\name{get_tf_name}
\alias{get_tf_name}
\title{Convert TF protein names into symbols}
\usage{
get_tf_name(list_tfs)
}
\arguments{
\item{list_tfs}{A character vector}
}
\value{
A character vector of same size as \code{list_tfs}
}
\description{
Convert TF protein names into symbols
}
