% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_info.R
\name{get_tf_symmetry}
\alias{get_tf_symmetry}
\title{Get symmetry info for a given TF or list of TFs}
\usage{
get_tf_symmetry(list_tfs)
}
\value{
A character vector of same size as \code{list_tfs}
}
\description{
Get symmetry info for a given TF or list of TFs
}
