% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_info.R
\name{get_tf_synonyms}
\alias{get_tf_synonyms}
\title{Get tf synonyms from the master table for a given vector of tfs}
\usage{
get_tf_synonyms(list_tfs)
}
\arguments{
\item{list_tfs}{A character vector of tf names}
}
\value{
A character vector of same size as \code{list_tfs} with comma-separated synonyms
}
\description{
Get tf synonyms from the master table for a given vector of tfs
}
