% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfs.R
\name{get_tfs_by_family}
\alias{get_tfs_by_family}
\title{Get TFs from a specific family}
\usage{
get_tfs_by_family(family)
}
\value{
A character vector of TF names
}
\description{
Get TFs from a specific family
}
