% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_upstream_gene}
\alias{get_upstream_gene}
\title{Get previous genes bnumbers for a given vector of genes, with respect to strand orientation}
\usage{
get_upstream_gene(list_genes)
}
\arguments{
\item{list_genes}{A character vector of valid bnumbers (previously checked/converted with convert_gene(to = "bnumber"))}
}
\value{
A vector of same size as \code{list_genes} (unknown input genes result in NA)
}
\description{
Get previous genes bnumbers for a given vector of genes, with respect to strand orientation
}
