% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_tf.R
\name{is_confirmed_tf}
\alias{is_confirmed_tf}
\title{Check if TF name is valid}
\usage{
is_confirmed_tf(tf_list)
}
\arguments{
\item{tf_list}{A character vector}
}
\value{
A logical vector
}
\description{
Check if TF name is valid
}
