% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_position.R
\name{is_genic}
\alias{is_genic}
\title{Check if a position or vector of positions are genic}
\usage{
is_genic(positions, strand)
}
\arguments{
\item{positions}{A vector of integers}

\item{strand}{A character vector of size 1 or same size as \code{positions}}
}
\value{
A vector of logicals of same size as \code{positions}
}
\description{
Check if a position or vector of positions are genic
}
