% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gene.R
\name{is_phantomgene}
\alias{is_phantomgene}
\title{Check if genes are phantom genes}
\usage{
is_phantomgene(bnum_list)
}
\arguments{
\item{bnum_list}{A character vector}
}
\value{
A logical vector
}
\description{
Check if genes are phantom genes
}
