% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_tf.R
\name{is_predicted_tf}
\alias{is_predicted_tf}
\title{Check if TF is predicted (from either Perez-Rueda 2015, Flores-Bautista 2020, Kim 2020)}
\usage{
is_predicted_tf(tf_list)
}
\arguments{
\item{tf_list}{A character vector}
}
\value{
A logical vector
}
\description{
Check if TF is predicted (from either Perez-Rueda 2015, Flores-Bautista 2020, Kim 2020)
}
