% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gene.R
\name{is_pseudo_gene}
\alias{is_pseudo_gene}
\title{Check if genes are pseudo genes}
\usage{
is_pseudo_gene(bnum_list)
}
\arguments{
\item{bnum_list}{A character vector}
}
\value{
A logical vector
}
\description{
Check if genes are pseudo genes
}
