% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_genes.R
\name{is_target_gene}
\alias{is_target_gene}
\title{Check if genes are TF-coding}
\usage{
is_target_gene(bnum_list)
}
\arguments{
\item{bnum_list}{A character vector}
}
\value{
A logical vector
}
\description{
Check if genes are TF-coding
}
