% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gene.R
\name{is_tf_gene}
\alias{is_tf_gene}
\title{Check if genes are TF-coding}
\usage{
is_tf_gene(bnum_list, source = "all")
}
\arguments{
\item{bnum_list}{A character vector}

\item{source}{Optional, one of c("all", "zika", "regulondb")}
}
\value{
A logical vector
}
\description{
Check if genes are TF-coding
}
