% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gene.R
\name{is_valid_bnumber}
\alias{is_valid_bnumber}
\title{Check if genes are valid bnumbers}
\usage{
is_valid_bnumber(bnum_list)
}
\arguments{
\item{bnum_list}{A character vector}
}
\value{
A logical vector
}
\description{
Check if genes are valid bnumbers
}
