% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_source_file.R
\name{read_coords_file}
\alias{read_coords_file}
\title{Read the file that contains genomic coordinates correspondence between 96.3 and 96.2}
\usage{
read_coords_file()
}
\value{
a data.frame
}
\description{
Read the file that contains genomic coordinates correspondence between 96.3 and 96.2
}
