% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_source_file.R
\name{read_master_tf_file}
\alias{read_master_tf_file}
\title{Read the reference tf file \link{obsolete}}
\usage{
read_master_tf_file()
}
\value{
a data.frame
}
\description{
Read the reference tf file \link{obsolete}
}
