% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_to_bnumber.R
\name{symbol_to_bnumber}
\alias{symbol_to_bnumber}
\title{Read the reference gene file}
\usage{
symbol_to_bnumber(list_symbols)
}
\arguments{
\item{list_symbols}{A character vector}
}
\value{
A character vector of same size as \code{list_symbols}
}
\description{
Read the reference gene file
}
