% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_to_symbol.R
\name{tf_to_symbol}
\alias{tf_to_symbol}
\title{Convert TF protein names into symbols}
\usage{
tf_to_symbol(list_proteins)
}
\arguments{
\item{list_proteins}{A character vector}
}
\value{
A character vector of same size as \code{list_proteins}
}
\description{
Convert TF protein names into symbols
}
