% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_position.R
\name{where_genic}
\alias{where_genic}
\title{Check in which gene falls a given position}
\usage{
where_genic(positions, strand)
}
\arguments{
\item{positions}{A vector of integers}

\item{strand}{A character vector of size 1 or same size as \code{positions}}
}
\value{
A character vector of same size as \code{positions}
}
\description{
Check in which gene falls a given position
}
