% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_position.R
\name{where_genic_rel}
\alias{where_genic_rel}
\title{Check where in gene}
\usage{
where_genic_rel(positions, strand)
}
\arguments{
\item{positions}{A numeric vector}

\item{strand}{A character vector}
}
\value{
A numeric vector of same size
}
\description{
Check where in gene
}
