% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_master_gene_file.R
\name{write_master_gene_file}
\alias{write_master_gene_file}
\title{Write the reference gene file}
\usage{
write_master_gene_file(data)
}
\value{
a data.frame
}
\description{
Write the reference gene file
}
