% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnumber_to_symbol.R
\name{bnumber_to_symbol}
\alias{bnumber_to_symbol}
\title{Read the reference gene file}
\usage{
bnumber_to_symbol(list_bnumbers)
}
\arguments{
\item{list_bnumbers}{A character vector}
}
\value{
A character vector of same size as \code{list_bnumbers}
}
\description{
Read the reference gene file
}
