% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_position.R
\name{closest_gene}
\alias{closest_gene}
\title{Get closest genes from genomic positions given a strand}
\usage{
closest_gene(positions, strand = "+-")
}
\arguments{
\item{positions}{A vector of integers}

\item{strand}{A character vector of size 1 or same size as \code{positions}}
}
\value{
A character vector of same size as \code{positions}
}
\description{
Get closest genes from genomic positions given a strand
}
