% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tf.R
\name{get_tf_bnumber}
\alias{get_tf_bnumber}
\title{Convert TF protein names into gene bnumbers \link{deprecated}}
\usage{
get_tf_bnumber(list_tfs)
}
\arguments{
\item{list_tfs}{A character vector}
}
\value{
A character vector of same size as \code{list_tfs}
}
\description{
Convert TF protein names into gene bnumbers \link{deprecated}
}
