% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_tf_genes}
\alias{get_tf_genes}
\title{Get TF-coding genes}
\usage{
get_tf_genes(source = "all")
}
\arguments{
\item{source}{Optional, one of c("all", "regulondb")}
}
\value{
A character vector
}
\description{
Get TF-coding genes
}
