% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tf.R
\name{get_tf_symbol}
\alias{get_tf_symbol}
\title{Convert TF protein names into symbols}
\usage{
get_tf_symbol(list_tfs)

get_tf_symbol(list_tfs)
}
\arguments{
\item{list_tfs}{A character vector}
}
\value{
A character vector of same size as \code{list_tfs}

A character vector of same size as \code{list_tfs}
}
\description{
Convert TF protein names into symbols

Convert TF protein names into symbols
}
