% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_position.R
\name{is_genic}
\alias{is_genic}
\title{Check if a position or vector of positions are genic}
\usage{
is_genic(positions)
}
\arguments{
\item{positions}{A numeric vector}
}
\value{
A vector of logicals of size \code{length(positions)}
}
\description{
Check if a position or vector of positions are genic
}
