% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gene.R
\name{is_valid_bnumber}
\alias{is_valid_bnumber}
\title{Check if bnumbers are valid}
\usage{
is_valid_bnumber(bnum_list)
}
\arguments{
\item{bnum_list}{A character vector}
}
\value{
A logical vector of same size as \code{bnum_list}
}
\description{
Check if bnumbers are valid
}
