% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_interval.R
\name{what_genes}
\alias{what_genes}
\title{Get bnumbers from genes entirely comprised in a given interval of coordinates}
\usage{
what_genes(start, stop, strand)
}
\arguments{
\item{start}{A numeric vector}

\item{stop}{A numeric vector}

\item{strand}{A character vector}
}
\value{
A vector of characters of same size as inputs
}
\description{
Get bnumbers from genes entirely comprised in a given interval of coordinates
}
