% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_position.R
\name{where_genic}
\alias{where_genic}
\title{Check in which gene}
\usage{
where_genic(positions)
}
\arguments{
\item{positions}{A numeric vector}
}
\value{
A character vector of same size
}
\description{
Check in which gene
}
